Fixed isSubscriptionsPage
Setting the subscriptions feed to list-view and refreshing the page caused the whole feed to be permanently hidden. This could only be fixed by disabling the script or navigating to https://www.youtube.com/feed/subscriptions?flow=1 from the search bar. Now, no visible changes should ever be made when in list view.